+++ tools/c_wrapper.cpp
@@ -236,7 +236,7 @@ const char *vrvToolkit_getTimesForElement(void *tkPtr, const char *xmlId)
     return tk->GetCString();
 }
 
-const char *vrvToolkit_getPitchPosition(void *tkPtr, double scoreTime, int midiPitch, int staff)
+const char *vrvToolkit_getPitchPosition(void *tkPtr, double scoreTime, double midiPitch, int staff)
 {
     Toolkit *tk = static_cast<Toolkit *>(tkPtr);
     tk->SetCString(tk->GetPitchPosition(scoreTime, midiPitch, staff));
